+++ src/D20Tek.BlazorComponents.Pager/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using D20Tek.BlazorComponents.Utilities;
+global using Microsoft.AspNetCore.Components;