+++ src/D20Tek.BlazorComponents.Vertically/GlobalUsings.cs
@@ -1,0 +1,3 @@
+global using D20Tek.BlazorComponents.Utilities;
+global using D20Tek.Vertically.Queries.Pagination;
+global using Microsoft.AspNetCore.Components;